Lazy-load longest axis length; fix longest axis length
The longest axis length algorithm before was actually calculating the
longest line length, which is not what the documentation indicated.
Furthermore, we were not actually using that value within the sharpmath
library, and I don't see it being regularly referenced within my projects,
so it seems reasonable to lazy-load this moderately expensive operation.

There might be a more clever way to do this using projections.